+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 * 0.1.0 (2023-11-04): Initial release
 * 0.1.1 (2023-11-04): Fix readme
 * 0.1.2 (2023-11-04): Add examples to readme
+* 0.2.0 (2023-11-05): Colorized Markdown output; added `-B`, `-C`, `-f`, `-r`
+  options; error on invalid target(s); update dependencies